// frontend/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        VeiculoEmOperacaoResponse: {
            /** Format: int64 */
            veiculoId: number;
            placa: string;
            placaFormatada: string;
            modelo: string;
            categoria: string;
            categoriaDescricao: string;
            exigeTesteFumacaPreta: boolean;
            possuiRastreador: boolean;
            possuiAdesivo: boolean;
            locadora: string;
            grupoTarifario?: string;
            /** Format: int64 */
            contratoId: number;
            obraCodigo: string;
            obraNome: string;
            localRetirada?: string;
            /** Format: int32 */
            pacoteKmContratado?: number;
            /** Format: date */
            dataRetirada?: string;
            /** Format: int64 */
            condutorId?: number;
            condutorNome?: string;
            condutorCargo?: string;
            /** Format: date */
            cnhValidade?: string;
            /** Format: int64 */
            diasParaVencerCnh?: number;
            cnhVencida: boolean;
            cnhEmAlerta: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}